Refactor: pull methods up
